+++ foxconn-sw-tool-business/src/main/java/com/foxconn/sw/business/tools/ToolsHistoryBusiness.java
@@ -37,7 +37,7 @@ public class ToolsHistoryBusiness {
         return toolDTOs;
     }
 
-    public Integer countByToolID(IntegerParams integerParams) {
+    public Long countByToolID(IntegerParams integerParams) {
         return swToolsHistoryExtensionMapper.countByToolID(integerParams.getParams());
     }
 }
