+++ foxconn-sw-tool-data/src/main/java/com/foxconn/sw/data/mapper/extension/system/SwPropertyExtensionMapper.java
@@ -97,5 +97,5 @@ public interface SwPropertyExtensionMapper extends SwPropertyMapper {
             "ORDER BY id ",
             "</script>",
     })
-    Integer getTotalCountByParams(PropertiesParams params);
+    Long getTotalCountByParams(PropertiesParams params);
 }
